homework done

+++ Library/Repositories/Repo.cs
@@ -19,7 +19,6 @@ namespace Library.Repositories
         DataSet set = new DataSet();
         public Repo()
         {
-            //conn = new SqlConnection();
             using (conn = new SqlConnection())
             {
                 var da = new SqlDataAdapter();
@@ -44,6 +43,7 @@ namespace Library.Repositories
             {
                 var command = new SqlCommand("INSERT INTO Authors(Id,FirstName,LastName) VALUES(@id,@firstName,@lastName)", conn);
                 conn.ConnectionString = cs;
+                conn.Open();
 
                 command.Parameters.Add(new SqlParameter
                 {
@@ -68,6 +68,7 @@ namespace Library.Repositories
 
                 var da = new SqlDataAdapter();
                 da.InsertCommand = command;
+                da.InsertCommand.ExecuteNonQuery();
                 da.Update(set, "AuthorsSet");
                 set.Clear();
 
@@ -79,25 +80,31 @@ namespace Library.Repositories
 
         public void DeleteAuthor(int id)
         {
-            using (conn = new SqlConnection())
+            if (id != -1)
             {
-                var command = new SqlCommand("DELETE FROM Authors WHERE Id=@id;", conn);
-                conn.ConnectionString = cs;
 
-                command.Parameters.Add(new SqlParameter
+                using (conn = new SqlConnection())
                 {
-                    DbType = DbType.Int32,
-                    ParameterName = "@id",
-                    Value = id
-                });
+                    conn.ConnectionString = cs;
+                    var command = new SqlCommand("DELETE FROM Authors WHERE Id=@id", conn);
+                    conn.Open();
 
-                var da = new SqlDataAdapter();
-                da.DeleteCommand = command;
-                da.Update(set, "AuthorsSet");
-                set.Clear();
+                    command.Parameters.Add(new SqlParameter
+                    {
+                        DbType = DbType.Int32,
+                        ParameterName = "@id",
+                        Value = id
+                    });
 
-                da = new SqlDataAdapter("SELECT * FROM Authors", conn);
-                da.Fill(set, "AuthorsSet");
+                    var da = new SqlDataAdapter();
+                    da.DeleteCommand = command;
+                    da.DeleteCommand.ExecuteNonQuery();
+                    da.Update(set, "AuthorsSet");
+                    set.Clear();
+
+                    da = new SqlDataAdapter("SELECT * FROM Authors", conn);
+                    da.Fill(set, "AuthorsSet");
+                }
             }
         }
     }
